+++ routes/index.tsx
@@ -1,18 +1,19 @@
-import { Header, Layout, Main, Text } from 'lunchbox';
-// import { ContributionCalendar } from "../components/ContributionCalendar/index.tsx";
+import { Header, Layout, Main, Panel, Text } from 'lunchbox';
+import { useSignal } from '@preact/signals';
 import { InputNote } from '../islands/InputNote/index.tsx';
+import { EntryList } from '../islands/EntryList/index.tsx';
 
 export default function Home() {
+  const updateLocalStorage = useSignal<number>(0);
+
   return (
-    <div>
-      <Header layout_type='full'>
-        <Text type='display'>Home</Text>
-      </Header>
-      <Main>
-        <Layout type='full'>
-          <InputNote />
-        </Layout>
-      </Main>
-    </div>
+    <Main>
+      <Layout type='full'>
+        <InputNote updateLocalStorage={updateLocalStorage} />
+      </Layout>
+      <Layout class='py-6' type='full'>
+        <EntryList updateLocalStorage={updateLocalStorage} />
+      </Layout>
+    </Main>
   );
 }